Rename OffenseMessageAnnotater to MessageAnnotator Rename MessageAnnotator#annotate_message to MessageAnnotator#annotate Add comments with yard doc syntax to public annotate method.
