+++ spec/rubocop/cop/message_annotator_spec.rb
@@ -2,19 +2,19 @@
 
 require 'spec_helper'
 
-describe RuboCop::Cop::OffenseMessageAnnotater do
+describe RuboCop::Cop::MessageAnnotator do
   let(:options) { {} }
   let(:config) { RuboCop::Config.new({}) }
   let(:formatter) { described_class.new(config, config['Cop/Cop'], options) }
 
-  describe '#annotate_message' do
-    subject(:annotate_message) do
-      formatter.annotate_message('message', 'Cop/Cop')
+  describe '#annotate' do
+    subject(:annotate) do
+      formatter.annotate('message', 'Cop/Cop')
     end
 
     context 'with default options' do
       it 'returns the message' do
-        expect(annotate_message).to eq('message')
+        expect(annotate).to eq('message')
       end
     end
 
@@ -36,20 +36,30 @@ describe RuboCop::Cop::OffenseMessageAnnotater do
       end
 
       it 'returns an annotated message' do
-        expect(annotate_message).to eq(
+        expect(annotate).to eq(
           'Cop/Cop: message my cop details (http://example.org/styleguide)'
         )
       end
     end
   end
 
-  describe '#style_guide_url' do
-    subject(:url) { formatter.style_guide_url }
+  describe 'with style guide url' do
+    subject(:annotate) do
+      formatter.annotate('', 'Cop/Cop')
+    end
+
+    let(:options) do
+      {
+        display_style_guide: true
+      }
+    end
 
     context 'when StyleGuide is not set in the config' do
       let(:config) { RuboCop::Config.new({}) }
 
-      it { is_expected.to be_nil }
+      it 'does not add style guide url' do
+        expect(annotate).to eq('')
+      end
     end
 
     context 'when StyleGuide is set in the config' do
@@ -59,7 +69,9 @@ describe RuboCop::Cop::OffenseMessageAnnotater do
         )
       end
 
-      it { is_expected.to eq('http://example.org/styleguide') }
+      it 'adds style guide url' do
+        expect(annotate).to include('http://example.org/styleguide')
+      end
     end
 
     context 'when a base URL is specified' do
@@ -73,17 +85,17 @@ describe RuboCop::Cop::OffenseMessageAnnotater do
 
       it 'does not specify a URL if a cop does not have one' do
         config['Cop/Cop'] = { 'StyleGuide' => nil }
-        expect(url).to be_nil
+        expect(annotate).to eq('')
       end
 
       it 'combines correctly with a target-based setting' do
         config['Cop/Cop'] = { 'StyleGuide' => '#target_based_url' }
-        expect(url).to eq('http://example.org/styleguide#target_based_url')
+        expect(annotate).to include('http://example.org/styleguide#target_based_url')
       end
 
       it 'can use a path-based setting' do
         config['Cop/Cop'] = { 'StyleGuide' => 'cop/path/rule#target_based_url' }
-        expect(url).to eq('http://example.org/cop/path/rule#target_based_url')
+        expect(annotate).to include('http://example.org/cop/path/rule#target_based_url')
       end
 
       it 'can accept relative paths if base has a full path' do
@@ -93,12 +105,12 @@ describe RuboCop::Cop::OffenseMessageAnnotater do
         config['Cop/Cop'] = {
           'StyleGuide' => '../rails-style-guide#target_based_url'
         }
-        expect(url).to eq('http://github.com/bbatsov/rails-style-guide#target_based_url')
+        expect(annotate).to include('http://github.com/bbatsov/rails-style-guide#target_based_url')
       end
 
       it 'allows absolute URLs in the cop config' do
         config['Cop/Cop'] = { 'StyleGuide' => 'http://other.org#absolute_url' }
-        expect(url).to eq('http://other.org#absolute_url')
+        expect(annotate).to include('http://other.org#absolute_url')
       end
     end
   end
